+++ app/config/config.yml
@@ -74,15 +74,6 @@ services:
         tags: [{ name: twig.extension }]
 
 surfnet_stepup_ra_ra:
-    sms:
-        originator: %sms_originator%
-        otp_expiry_interval: %sms_otp_expiry_interval%
-        maximum_otp_requests: %sms_maximum_otp_requests%
-    gateway_api:
-        url: %gateway_api_url%
-        credentials:
-            username: %gateway_api_username%
-            password: %gateway_api_password%
     required_loa: %loa_required_for_login%
 
 mopa_bootstrap:
@@ -125,6 +116,15 @@ surfnet_stepup:
         loa1: %stepup_loa_loa1%
         loa2: %stepup_loa_loa2%
         loa3: %stepup_loa_loa3%
+    sms:
+        originator: %sms_originator%
+        otp_expiry_interval: %sms_otp_expiry_interval%
+        maximum_otp_requests: %sms_maximum_otp_requests%
+    gateway_api:
+        url: %gateway_api_url%
+        credentials:
+            username: %gateway_api_username%
+            password: %gateway_api_password%
 
 jms_translation:
     locales: %locales%
