+++ src/Surfnet/StepupRa/RaBundle/DependencyInjection/Configuration.php
@@ -32,62 +32,15 @@ class Configuration implements ConfigurationInterface
         $treeBuilder = new TreeBuilder();
         $rootNode = $treeBuilder->root('surfnet_stepup_ra_ra');
 
-        $this->createGatewayApiConfiguration($rootNode);
-        $this->createSmsConfiguration($rootNode);
+        $this->createRaConfiguration($rootNode);
 
         return $treeBuilder;
     }
 
-    private function createGatewayApiConfiguration(ArrayNodeDefinition $root)
+    private function createRaConfiguration(ArrayNodeDefinition $root)
     {
         $root
             ->children()
-                ->arrayNode('gateway_api')
-                    ->info('Gateway API configuration')
-                    ->children()
-                        ->arrayNode('credentials')
-                            ->info('Basic authentication credentials')
-                            ->children()
-                                ->scalarNode('username')
-                                    ->info('Username for the Gateway API')
-                                    ->isRequired()
-                                    ->validate()
-                                        ->ifTrue(function ($value) {
-                                            return (!is_string($value) || empty($value));
-                                        })
-                                        ->thenInvalid(
-                                            'Invalid Gateway API username specified: "%s". Must be non-empty string'
-                                        )
-                                    ->end()
-                                ->end()
-                                ->scalarNode('password')
-                                    ->info('Password for the Gateway API')
-                                    ->isRequired()
-                                    ->validate()
-                                        ->ifTrue(function ($value) {
-                                            return (!is_string($value) || empty($value));
-                                        })
-                                        ->thenInvalid(
-                                            'Invalid Gateway API password specified: "%s". Must be non-empty string'
-                                        )
-                                    ->end()
-                                ->end()
-                            ->end()
-                        ->end()
-                        ->scalarNode('url')
-                            ->info('The URL to the Gateway application (e.g. https://gateway.tld)')
-                            ->isRequired()
-                            ->validate()
-                                ->ifTrue(function ($value) {
-                                    return (!is_string($value) || empty($value) || !preg_match('~/$~', $value));
-                                })
-                                ->thenInvalid(
-                                    'Invalid Gateway URL specified: "%s". Must be string ending in forward slash'
-                                )
-                            ->end()
-                        ->end()
-                    ->end()
-                ->end()
                 ->scalarNode('required_loa')
                     ->info('The required LOA to be able to log in, should match the loa defined at the gateway')
                     ->isRequired()
@@ -96,56 +49,6 @@ class Configuration implements ConfigurationInterface
                         return !is_string($value);
                     })
                         ->thenInvalid('the required loa must be a string')
-                    ->end()
-                ->end()
-            ->end();
-    }
-
-    private function createSmsConfiguration(ArrayNodeDefinition $root)
-    {
-        $root
-            ->children()
-                ->arrayNode('sms')
-                    ->info('SMS configuration')
-                    ->isRequired()
-                    ->children()
-                        ->scalarNode('originator')
-                            ->info('Originator (sender) for SMS messages')
-                            ->isRequired()
-                            ->validate()
-                                ->ifTrue(function ($value) {
-                                    return (!is_string($value) || !preg_match('~^[a-z0-9]{1,11}$~i', $value));
-                                })
-                                ->thenInvalid(
-                                    'Invalid SMS originator specified: "%s". Must be a string matching '
-                                    . '"~^[a-z0-9]{1,11}$~i".'
-                                )
-                            ->end()
-                        ->end()
-                        ->integerNode('otp_expiry_interval')
-                            ->info('After how many seconds an SMS challenge OTP expires')
-                            ->isRequired()
-                            ->validate()
-                                ->ifTrue(function ($value) {
-                                    return $value <= 0;
-                                })
-                                ->thenInvalid(
-                                    'Invalid SMS challenge OTP expiry, must be one or more seconds.'
-                                )
-                            ->end()
-                        ->end()
-                        ->integerNode('maximum_otp_requests')
-                            ->info('How many challenges a user may request during a session')
-                            ->isRequired()
-                            ->validate()
-                                ->ifTrue(function ($value) {
-                                    return $value <= 0;
-                                })
-                                ->thenInvalid(
-                                    'Maximum OTP requests has a minimum of 1'
-                                )
-                            ->end()
-                        ->end()
                     ->end()
                 ->end()
             ->end();
