+++ src/Surfnet/StepupRa/RaBundle/Form/Type/SendSmsChallengeType.php
@@ -35,7 +35,7 @@ class SendSmsChallengeType extends AbstractType
     public function setDefaultOptions(OptionsResolverInterface $resolver)
     {
         $resolver->setDefaults([
-            'data_class' => 'Surfnet\StepupRa\RaBundle\Command\SendSmsChallengeCommand',
+            'data_class' => 'Surfnet\StepupBundle\Command\SendSmsChallengeCommand',
         ]);
     }
