+++ src/Surfnet/StepupRa/RaBundle/Security/Authentication/SessionHandler.php
@@ -19,7 +19,6 @@
 namespace Surfnet\StepupRa\RaBundle\Security\Authentication;
 
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
-use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 
 class SessionHandler
 {
